Description: Component Callback logic as per discussed flow dialgram. Bug ID: US7957 Reviewer: Ritesh Jha Comments: NA

+++ Source/Library/multiproduct/src/main/java/com/philips/multiproduct/listfragment/ProductListingFragment.java
@@ -188,6 +188,7 @@ public class ProductListingFragment extends MultiProductBaseFragment {
     @Override
     public void onDestroy() {
         super.onDestroy();
-        mProductDetailsListener.deleteObserver(mDetailedScreenFragment);
+        if (mProductDetailsListener != null)
+            mProductDetailsListener.deleteObserver(mDetailedScreenFragment);
     }
 }
